splitting models into a package
